*: fix GCC warnings, #24
- enable -Wall by default, it's good (-Werror purposely NOT added)
- add some casts
- inject nbsp for NoBreak

Closes: https://github.com/grobian/html2text/issues/24

+++ Area.cpp
@@ -108,7 +108,7 @@ Line::Line(const istr &s):
 	cells_(malloc_array(Cell, length_))
 {
 	Cell *q = cells_;
-	for (int i = 0; i < length_; i++) {
+	for (size_t i = 0; i < (size_t)length_; i++) {
 		q->character = s[i];
 		q->attribute = Cell::NONE;
 		q++;
